change bed monitor provider
